requirements.txt
botbuilder-core~=4.14.1
botbuilder-dialogs~=4.14.1
botbuilder-ai~=4.14.1
aiohttp~=3.6.2
spacy~=2.3.2
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-2.3.1/en_core_web_lg-2.3.1.tar.gz
pandas~=1.1.1
pm4py~=1.4.1.1
ppinot4py~=1.2
PyYAML~=6.0
pytest~=6.2.2
numpy==1.22.3
pathlib==1.0.1
requests~=2.25.0
ruamel.yaml~=0.16
colorama~=0.4.4
fastDamerauLevenshtein~=1.0.7
pyjarowinkler~=1.8
openpyxl~=3.0.9
torch
torchvision
transformers
sentence-transformers
sqlalchemy
